Remove all instances of timer (first round)
My first run through the different files and removing the instances of the timer. There is a section in main.py that I have commented out because it is a try/except/finally chain that relies on the timer and I am not sure which parts of it need to be saved.